requirements.txt
pyyaml
requests
dnspython